+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   is included][14028]
 - [Function docs in autocomplete in table expressions][14059]
 - [Many CLI arguments removed][14069]
+- [Dragging edges from plus button on nodes is now possible][14246]
 - [JSON and SQL visualizations' content may be now selected and copied][14262]
 - [SQL visualization displays interpolated parameters properly][14262]
 - [Added immediate rename of user created components][14209]
@@ -32,6 +33,7 @@
 [14059]: https://github.com/enso-org/enso/pull/14059
 [14069]: https://github.com/enso-org/enso/pull/14069
 [14262]: https://github.com/enso-org/enso/pull/14262
+[14246]: https://github.com/enso-org/enso/pull/14246
 [14209]: https://github.com/enso-org/enso/pull/14209
 [14215]: https://github.com/enso-org/enso/pull/14215
 [14270]: https://github.com/enso-org/enso/pull/14270